Refactored utilFns
