+++ simplq/src/components/pages/Home/CreateJoinForm.jsx
@@ -2,7 +2,8 @@ import React, { useState } from 'react';
 import { useHistory } from 'react-router';
 import * as QueueService from '../../../services/queue';
 import styles from './home.module.scss';
-import { handleEnterPress, isQueueNameValid } from '../../common/utilFns';
+import { handleEnterPress } from '../../../utils/eventHandling';
+import { isQueueNameValid } from '../../../utils/textOperations';
 import InputField from '../../common/InputField';
 import LoadingIndicator from '../../common/LoadingIndicator';
 import StandardButton from '../../common/Button';
