+++ simplq/src/components/pages/Home/index.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { BenefitsInfo, HowToCreate, HowToJoin, ExtraInfo, GetStarted } from './StaticInfos';
 import LandingPage from './LandingPage';
-import { smoothScrollToHomePageTop } from '../../common/utilFns';
+import { smoothScrollToHomePageTop } from '../../../utils/scrollingOperations';
 
 const Home = () => {
   // Scroll to top whenever home page is mounted
