+++ simplq/src/components/pages/PageNotFound/pageNotFound.module.scss
@@ -20,7 +20,7 @@
     padding: 0 40px;
   }
   .buttondiv {
-    display: flex !important;
+    display: flex;
     justify-content: center;
   }
   .errorimg {
