+++ simplq/src/utils/eventHandling.js
@@ -1,0 +1,7 @@
+/* eslint-disable import/prefer-default-export */
+
+export const handleEnterPress = (event, keyPresshandler) => {
+  if (event.key === 'Enter') {
+    keyPresshandler();
+  }
+};
