+++ simplq/src/utils/scrollingOperations.js
@@ -1,11 +1,3 @@
-export const handleEnterPress = (event, keyPresshandler) => {
-  if (event.key === 'Enter') {
-    keyPresshandler();
-  }
-};
-
-export const isQueueNameValid = (qname) => qname.match('^[A-Za-z0-9-]*$'); // Only letters, numbers and - allowed
-
 export const smoothScrollTo = (targetElement) => {
   // offset to avoid the sticky header on top to block the "Create a Queue at the click of a button" text
   const offset = 45;
@@ -29,6 +21,3 @@ export const smoothScrollToHomePageTop = () => {
     window.location.href = '/';
   }
 };
-
-export const getSentenceCaseText = (txt) =>
-  txt?.charAt(0)?.toUpperCase() + txt?.substr(1)?.toLowerCase();
